tools/nemo_runner/pyproject.toml
[project]
name = "stt-bench-matrix-nemo-runner"
version = "0.1.0"
description = "Isolated NeMo runner for STT benchmarks"
requires-python = ">=3.11"
dependencies = [
  "nemo_toolkit[asr]",
  "torch",
  "pyarrow<15",
  "datasets<3",
  "numpy<2",
]
